refact str to const
