+++ main.go
@@ -11,7 +11,6 @@ import (
 func main() {
 	goscout := app.New()
 	fyneWindow := goscout.NewWindow("GoScout")
-
 	cfg, err := ui.LoadConfig()
 	if err != nil {
 		log.Fatalf("Failed to load config: %v", err)
